+++ packages/product/src/services/product-module-service.ts
@@ -25,24 +25,48 @@ import {
   InternalModuleDeclaration,
   JoinerServiceConfig,
   ProductTypes,
+  IEventBusModuleService,
 } from "@medusajs/types"
-import { serialize } from "@mikro-orm/core"
 
 import ProductImageService from "./product-image"
-import { ProductServiceTypes, ProductVariantServiceTypes } from "../types/services"
+
+import {
+  CreateProductCategoryDTO,
+  ProductCategoryEventData,
+  ProductCategoryEvents,
+  UpdateProductCategoryDTO,
+} from "../types/services/product-category"
+
+import { UpdateProductVariantDTO } from "../types/services/product-variant"
+
+import {
+  ProductCollectionEventData,
+  ProductCollectionEvents,
+} from "../types/services/product-collection"
+
+import {
+  ProductEventData,
+  ProductEvents,
+  UpdateProductDTO,
+} from "../types/services/product"
+
 import {
   InjectManager,
   InjectTransactionManager,
   isDefined,
   isString,
   kebabCase,
+  mapObjectTo,
   MedusaContext,
   MedusaError,
 } from "@medusajs/utils"
 
 import { shouldForceTransaction } from "../utils"
-import { joinerConfig } from "./../joiner-config"
-import { ProductCategoryServiceTypes } from "../types"
+import {
+  entityNameToLinkableKeysMap,
+  joinerConfig,
+  LinkableKeys,
+} from "./../joiner-config"
 
 type InjectedDependencies = {
   baseRepository: DAL.RepositoryService
@@ -54,6 +78,7 @@ type InjectedDependencies = {
   productImageService: ProductImageService<any>
   productTypeService: ProductTypeService<any>
   productOptionService: ProductOptionService<any>
+  eventBusModuleService?: IEventBusModuleService
 }
 
 export default class ProductModuleService<
@@ -73,12 +98,16 @@ export default class ProductModuleService<
     TProductVariant,
     TProduct
   >
+
+  // eslint-disable-next-line max-len
   protected readonly productCategoryService_: ProductCategoryService<TProductCategory>
   protected readonly productTagService_: ProductTagService<TProductTag>
+  // eslint-disable-next-line max-len
   protected readonly productCollectionService_: ProductCollectionService<TProductCollection>
   protected readonly productImageService_: ProductImageService<TProductImage>
   protected readonly productTypeService_: ProductTypeService<TProductType>
   protected readonly productOptionService_: ProductOptionService<TProductOption>
+  protected readonly eventBusModuleService_?: IEventBusModuleService
 
   constructor(
     {
@@ -91,6 +120,7 @@ export default class ProductModuleService<
       productImageService,
       productTypeService,
       productOptionService,
+      eventBusModuleService,
     }: InjectedDependencies,
     protected readonly moduleDeclaration: InternalModuleDeclaration
   ) {
@@ -103,6 +133,7 @@ export default class ProductModuleService<
     this.productImageService_ = productImageService
     this.productTypeService_ = productTypeService
     this.productOptionService_ = productOptionService
+    this.eventBusModuleService_ = eventBusModuleService
   }
 
   __joinerConfig(): JoinerServiceConfig {
@@ -393,11 +424,12 @@ export default class ProductModuleService<
     config: FindConfig<ProductTypes.ProductOptionDTO> = {},
     @MedusaContext() sharedContext: Context = {}
   ): Promise<[ProductTypes.ProductOptionDTO[], number]> {
-    const [productOptions, count] = await this.productOptionService_.listAndCount(
-      filters,
-      config,
-      sharedContext
-    )
+    const [productOptions, count] =
+      await this.productOptionService_.listAndCount(
+        filters,
+        config,
+        sharedContext
+      )
 
     return [JSON.parse(JSON.stringify(productOptions)), count]
   }
@@ -491,6 +523,13 @@ export default class ProductModuleService<
       sharedContext
     )
 
+    await this.eventBusModuleService_?.emit<ProductCollectionEventData>(
+      productCollections.map(({ id }) => ({
+        eventName: ProductCollectionEvents.COLLECTION_CREATED,
+        data: { id },
+      }))
+    )
+
     return JSON.parse(JSON.stringify(productCollections))
   }
 
@@ -504,6 +543,13 @@ export default class ProductModuleService<
       sharedContext
     )
 
+    await this.eventBusModuleService_?.emit<ProductCollectionEventData>(
+      productCollections.map(({ id }) => ({
+        eventName: ProductCollectionEvents.COLLECTION_UPDATED,
+        data: { id },
+      }))
+    )
+
     return JSON.parse(JSON.stringify(productCollections))
   }
 
@@ -515,6 +561,13 @@ export default class ProductModuleService<
     await this.productCollectionService_.delete(
       productCollectionIds,
       sharedContext
+    )
+
+    await this.eventBusModuleService_?.emit<ProductCollectionEventData>(
+      productCollectionIds.map((id) => ({
+        eventName: ProductCollectionEvents.COLLECTION_DELETED,
+        data: { id },
+      }))
     )
   }
 
@@ -550,12 +603,17 @@ export default class ProductModuleService<
 
   @InjectTransactionManager(shouldForceTransaction, "baseRepository_")
   async createCategory(
-    data: ProductCategoryServiceTypes.CreateProductCategoryDTO,
+    data: CreateProductCategoryDTO,
     @MedusaContext() sharedContext: Context = {}
   ) {
     const productCategory = await this.productCategoryService_.create(
       data,
       sharedContext
+    )
+
+    await this.eventBusModuleService_?.emit<ProductCategoryEventData>(
+      ProductCategoryEvents.CATEGORY_CREATED,
+      { id: productCategory.id }
     )
 
     return JSON.parse(JSON.stringify(productCategory))
@@ -564,13 +622,18 @@ export default class ProductModuleService<
   @InjectTransactionManager(shouldForceTransaction, "baseRepository_")
   async updateCategory(
     categoryId: string,
-    data: ProductCategoryServiceTypes.UpdateProductCategoryDTO,
+    data: UpdateProductCategoryDTO,
     @MedusaContext() sharedContext: Context = {}
   ) {
     const productCategory = await this.productCategoryService_.update(
       categoryId,
       data,
       sharedContext
+    )
+
+    await this.eventBusModuleService_?.emit<ProductCategoryEventData>(
+      ProductCategoryEvents.CATEGORY_UPDATED,
+      { id: productCategory.id }
     )
 
     return JSON.parse(JSON.stringify(productCategory))
@@ -582,6 +645,11 @@ export default class ProductModuleService<
     @MedusaContext() sharedContext: Context = {}
   ): Promise<void> {
     await this.productCategoryService_.delete(categoryId, sharedContext)
+
+    await this.eventBusModuleService_?.emit<ProductCategoryEventData>(
+      ProductCategoryEvents.CATEGORY_DELETED,
+      { id: categoryId }
+    )
   }
 
   @InjectManager("baseRepository_")
@@ -599,12 +667,25 @@ export default class ProductModuleService<
     return JSON.parse(JSON.stringify(categories))
   }
 
-  async create(data: ProductTypes.CreateProductDTO[], sharedContext?: Context): Promise<ProductTypes.ProductDTO[]> {
+  async create(
+    data: ProductTypes.CreateProductDTO[],
+    sharedContext?: Context
+  ): Promise<ProductTypes.ProductDTO[]> {
     const products = await this.create_(data, sharedContext)
-
-    return this.baseRepository_.serialize<ProductTypes.ProductDTO[]>(products, {
+    const createdProducts = await this.baseRepository_.serialize<
+      ProductTypes.ProductDTO[]
+    >(products, {
       populate: true,
     })
+
+    await this.eventBusModuleService_?.emit<ProductEventData>(
+      createdProducts.map(({ id }) => ({
+        eventName: ProductEvents.PRODUCT_CREATED,
+        data: { id },
+      }))
+    )
+
+    return createdProducts
   }
 
   async update(
@@ -613,11 +694,20 @@ export default class ProductModuleService<
   ): Promise<ProductTypes.ProductDTO[]> {
     const products = await this.update_(data, sharedContext)
 
-    return this.baseRepository_.serialize<
+    const updatedProducts = await this.baseRepository_.serialize<
       ProductTypes.ProductDTO[]
     >(products, {
       populate: true,
     })
+
+    await this.eventBusModuleService_?.emit<ProductEventData>(
+      updatedProducts.map(({ id }) => ({
+        eventName: ProductEvents.PRODUCT_UPDATED,
+        data: { id },
+      }))
+    )
+
+    return updatedProducts
   }
 
   @InjectTransactionManager(shouldForceTransaction, "baseRepository_")
@@ -659,9 +749,18 @@ export default class ProductModuleService<
           productData.discountable = false
         }
 
-        await this.upsertAndAssignImagesToProductData(productData, sharedContext)
-        await this.upsertAndAssignProductTagsToProductData(productData, sharedContext)
-        await this.upsertAndAssignProductTypeToProductData(productData, sharedContext)
+        await this.upsertAndAssignImagesToProductData(
+          productData,
+          sharedContext
+        )
+        await this.upsertAndAssignProductTagsToProductData(
+          productData,
+          sharedContext
+        )
+        await this.upsertAndAssignProductTypeToProductData(
+          productData,
+          sharedContext
+        )
 
         return productData as CreateProductOnlyDTO
       })
@@ -679,9 +778,11 @@ export default class ProductModuleService<
     const productOptionsData = [...productOptionsMap]
       .map(([handle, options]) => {
         return options.map((option) => {
-          const productOptionsData: ProductTypes.CreateProductOptionOnlyDTO = { ...option }
+          const productOptionsData: ProductTypes.CreateProductOptionOnlyDTO = {
+            ...option,
+          }
           const product = productByHandleMap.get(handle)
-          const productId = product?.id!
+          const productId = product?.id
 
           if (productId) {
             productOptionsData.product_id = productId
@@ -729,31 +830,29 @@ export default class ProductModuleService<
     data: ProductTypes.UpdateProductDTO[],
     @MedusaContext() sharedContext: Context = {}
   ): Promise<TProduct[]> {
-    const productIds = data.map(pd => pd.id)
+    const productIds = data.map((pd) => pd.id)
     const existingProductVariants = await this.productVariantService_.list(
       { product_id: productIds },
       {},
       sharedContext
     )
 
-    const existingProductVariantsMap = new Map<
-      string,
-      ProductVariant[]
-    >(
+    const existingProductVariantsMap = new Map<string, ProductVariant[]>(
       data.map((productData) => {
-        const productVariantsForProduct = existingProductVariants
-          .filter((variant) => variant.product_id === productData.id)
+        const productVariantsForProduct = existingProductVariants.filter(
+          (variant) => variant.product_id === productData.id
+        )
 
-        return [
-          productData.id,
-          productVariantsForProduct,
-        ]
+        return [productData.id, productVariantsForProduct]
       })
     )
 
     const productVariantsMap = new Map<
       string,
-      (ProductTypes.CreateProductVariantDTO | ProductTypes.UpdateProductVariantDTO)[]
+      (
+        | ProductTypes.CreateProductVariantDTO
+        | ProductTypes.UpdateProductVariantDTO
+      )[]
     >()
 
     const productOptionsMap = new Map<
@@ -779,11 +878,20 @@ export default class ProductModuleService<
           productData.discountable = false
         }
 
-        await this.upsertAndAssignImagesToProductData(productData, sharedContext)
-        await this.upsertAndAssignProductTagsToProductData(productData, sharedContext)
-        await this.upsertAndAssignProductTypeToProductData(productData, sharedContext)
+        await this.upsertAndAssignImagesToProductData(
+          productData,
+          sharedContext
+        )
+        await this.upsertAndAssignProductTagsToProductData(
+          productData,
+          sharedContext
+        )
+        await this.upsertAndAssignProductTypeToProductData(
+          productData,
+          sharedContext
+        )
 
-        return productData as ProductServiceTypes.UpdateProductDTO
+        return productData as UpdateProductDTO
       })
     )
 
@@ -797,10 +905,12 @@ export default class ProductModuleService<
     )
 
     const productOptionsData = [...productOptionsMap]
-      .map(([id, options]) => options.map((option) => ({
-        ...option,
-        product: productByIdMap.get(id)!,
-      })))
+      .map(([id, options]) =>
+        options.map((option) => ({
+          ...option,
+          product: productByIdMap.get(id)!,
+        }))
+      )
       .flat()
 
     const productOptions = await this.productOptionService_.create(
@@ -825,7 +935,7 @@ export default class ProductModuleService<
       const existingVariants = existingProductVariantsMap.get(productId)
 
       variants.forEach((variant) => {
-        const isVariantIdDefined = ("id" in variant) && isDefined(variant.id)
+        const isVariantIdDefined = "id" in variant && isDefined(variant.id)
 
         if (isVariantIdDefined) {
           variantsToUpdate.push(variant as ProductTypes.UpdateProductVariantDTO)
@@ -849,13 +959,13 @@ export default class ProductModuleService<
       productVariantsToCreateMap.set(productId, variantsToCreate)
       productVariantsToUpdateMap.set(productId, variantsToUpdate)
 
-      const variantsToUpdateIds = variantsToUpdate.map(v => v?.id) as string[]
-      const existingVariantIds = existingVariants?.map(v => v.id) || []
+      const variantsToUpdateIds = variantsToUpdate.map((v) => v?.id) as string[]
+      const existingVariantIds = existingVariants?.map((v) => v.id) || []
       const variantsToUpdateSet = new Set(variantsToUpdateIds)
 
       productVariantIdsToDelete.push(
         ...new Set(
-          existingVariantIds.filter(x => !variantsToUpdateSet.has(x))
+          existingVariantIds.filter((x) => !variantsToUpdateSet.has(x))
         )
       )
     }
@@ -876,7 +986,7 @@ export default class ProductModuleService<
       promises.push(
         this.productVariantService_.update(
           productByIdMap.get(productId)!,
-          variants as unknown as ProductVariantServiceTypes.UpdateProductVariantDTO[],
+          variants as unknown as UpdateProductVariantDTO[],
           sharedContext
         )
       )
@@ -884,7 +994,10 @@ export default class ProductModuleService<
 
     if (productVariantIdsToDelete.length) {
       promises.push(
-        this.productVariantService_.delete(productVariantIdsToDelete, sharedContext)
+        this.productVariantService_.delete(
+          productVariantIdsToDelete,
+          sharedContext
+        )
       )
     }
 
@@ -905,9 +1018,13 @@ export default class ProductModuleService<
 
     if (productData.images?.length) {
       productData.images = await this.productImageService_.upsert(
-        productData.images.map((image) =>
-          isString(image) ? image : image.url
-        ),
+        productData.images.map((image) => {
+          if (isString(image)) {
+            return image
+          } else {
+            return image.url
+          }
+        }),
         sharedContext
       )
     }
@@ -930,14 +1047,12 @@ export default class ProductModuleService<
     sharedContext: Context = {}
   ) {
     if (isDefined(productData.type)) {
-      const productType = (
-        await this.productTypeService_.upsert(
-          [productData.type],
-          sharedContext
-        )
+      const productType = await this.productTypeService_.upsert(
+        [productData.type!],
+        sharedContext
       )
 
-      productData.type_id = productType?.[0]?.id
+      productData.type = productType?.[0] as ProductTypes.CreateProductTypeDTO
     }
   }
 
@@ -947,24 +1062,63 @@ export default class ProductModuleService<
     @MedusaContext() sharedContext: Context = {}
   ): Promise<void> {
     await this.productService_.delete(productIds, sharedContext)
+
+    await this.eventBusModuleService_?.emit<ProductEventData>(
+      productIds.map((id) => ({
+        eventName: ProductEvents.PRODUCT_DELETED,
+        data: { id },
+      }))
+    )
   }
 
-  async softDelete(
+  async softDelete<
+    TReturnableLinkableKeys extends string = Lowercase<
+      keyof typeof LinkableKeys
+    >
+  >(
     productIds: string[],
+    {
+      returnLinkableKeys,
+    }: { returnLinkableKeys?: TReturnableLinkableKeys[] } = {
+      returnLinkableKeys: [],
+    },
     sharedContext: Context = {}
-  ): Promise<ProductTypes.ProductDTO[]> {
-    const products = await this.softDelete_(productIds, sharedContext)
+  ): Promise<Record<Lowercase<keyof typeof LinkableKeys>, string[]> | void> {
+    let [products, cascadedEntitiesMap] = await this.softDelete_(
+      productIds,
+      sharedContext
+    )
 
-    return this.baseRepository_.serialize<ProductTypes.ProductDTO[]>(products, {
+    const softDeletedProducts = await this.baseRepository_.serialize<
+      ProductTypes.ProductDTO[]
+    >(products, {
       populate: true,
     })
+
+    await this.eventBusModuleService_?.emit<ProductEventData>(
+      softDeletedProducts.map(({ id }) => ({
+        eventName: ProductEvents.PRODUCT_DELETED,
+        data: { id },
+      }))
+    )
+
+    let mappedCascadedEntitiesMap
+    if (returnLinkableKeys) {
+      mappedCascadedEntitiesMap = mapObjectTo<
+        Record<Lowercase<keyof typeof LinkableKeys>, string[]>
+      >(cascadedEntitiesMap, entityNameToLinkableKeysMap, {
+        pick: returnLinkableKeys as string[],
+      })
+    }
+
+    return mappedCascadedEntitiesMap ? mappedCascadedEntitiesMap : void 0
   }
 
   @InjectTransactionManager(shouldForceTransaction, "baseRepository_")
   protected async softDelete_(
     productIds: string[],
     @MedusaContext() sharedContext: Context = {}
-  ): Promise<TProduct[]> {
+  ): Promise<[TProduct[], Record<string, unknown[]>]> {
     return await this.productService_.softDelete(productIds, sharedContext)
   }