+++ packages/types/src/event-bus/event-bus-module.ts
@@ -4,7 +4,7 @@ export interface IEventBusModuleService {
   emit<T>(
     eventName: string,
     data: T,
-    options: Record<string, unknown>
+    options?: Record<string, unknown>
   ): Promise<void>
   emit<T>(data: EmitData<T>[]): Promise<void>